requirements.txt
# Core
numpy>=1.24
pandas>=2.0
openpyxl>=3.1

# Machine Learning
scikit-learn>=1.3
torch>=2.0
xgboost>=2.0  # optional, scikit-learn GradientBoosting used by default

# Time Series Forecasting
chronos-forecasting>=1.0  # Amazon Chronos (amazon/chronos-t5-small)
statsforecast>=1.5        # AutoARIMA, AutoETS

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Statistics
scipy>=1.11

# Data formats
pyreadstat>=1.2  # For reading .xpt (NHANES SAS transport files)
